graham: Fix a crash that occurs when loading the "Scissors lift mechanism" example
The graham scan algorithm require all points to be sorted regarding a polar
order. The polar order is defined by the 'Compare' function that consider
the colinear case as 'do nothing'.
In the graham function, if a,b and c are collinear, the triangle area is 0
and the point c is not considered on the Left side of the segment [a,b],
breaking the polar order. Later, the graham function tries to pop an element
and crash.
To prevent the crash, the 'Compare' function and 'graham' functions are now
consistent wrt the collinearity.